refactor: split example tests into single purpose modules
